backend/requirements.txt
requests
beautifulsoup4
fake-useragent
playwright
supabase
python-dotenv

